+++ NEWS.md
@@ -1,12 +1,16 @@
 # News about the `tcgsaseq` R package
 
-### Main changes in Version 1.6.4 (2017-11-28) --- *this is only a minor release*:
+
+### Main changes in Version 1.6.5 (2017-12-24) --- *this is only a minor release*:
 * bug fix in the log2-cpm transformation computation (for `preprocessed = FALSE`) in `sp_weights`
+* bug fix in `sp_weights` when facing `NA` or `NaN`values
+
 
 ### Main changes in Version 1.6.3 (2017-11-09)
 * `WARNING` instead of `ERROR` when a gene set with no measured genes is tested
 * `NA` support through `na.rm_...` logical arguments
 * bug fix for row.names error in weights computation for gene-wise analysis
+
 
 ### Main changes in Version 1.6.2 (2017-08-04)
 * bug fix to the log2-cpm transformation computation (for `preprocessed = FALSE`)